requirements.txt
fastapi
uvicorn
python-dotenv
PyGithub
psycopg2-binary
requests
langchain
langchain-anthropic
langchain-community
langgraph
chromadb
sentence-transformers
tree-sitter==0.21.3
tree-sitter-python==0.21.0
anthropic
pyngrok
plotly
pandas
